requirements.txt
gradio
transformers
torch
openai
python-dotenv
